fix: enhance AccessControlService, GameService, CoreService, GameRoleService

+++ src/domain/service/game_role.py
@@ -1,11 +1,6 @@
 from uuid import UUID
 
 from src.domain.exceptions import NotFoundException, ForbiddenException, InternalLogicException
-from src.domain.models.game_roles.request import (
-    GameRoleItemCreateRequest,
-    GameRoleItemUpdateRequest,
-    GameRoleSetUpdateRequest,
-)
 from src.infra.postgre.models import GameRoleSet, GameRole
 from src.infra.postgre.repo import (
     GameRoleSetRepository,
@@ -18,10 +13,10 @@ from src.infra.postgre import IntegrityForeignException, IntegrityUnknownExcepti
 
 class GameRoleService:
     def __init__(
-        self,
-        server_repo: ServerRepository,
-        role_set_repo: GameRoleSetRepository,
-        role_repo: GameRoleRepository,
+            self,
+            server_repo: ServerRepository,
+            role_set_repo: GameRoleSetRepository,
+            role_repo: GameRoleRepository,
     ) -> None:
         self.server_repo = server_repo
         self.role_set_repo = role_set_repo
@@ -34,44 +29,48 @@ class GameRoleService:
         return server.role_set
 
     async def update_role_set(
-        self, role_set_id: UUID, body: GameRoleSetUpdateRequest, permission_mask: int = 0 # TODO : Replace body
+            self, role_set_id: UUID, server_id: UUID, name: str | None = None, permission_mask: int = 0
     ) -> GameRoleSet:
-        # TODO: check if belongs to issuer member server
         if not PERMISSION.check_permission(permission_mask, PERMISSION.EDIT_ROLE_SET):
             raise ForbiddenException("Unable to edit role set")
-        role_set_field = await self.role_set_repo.get_by_id(role_set_id)
-        if role_set_field is None:
-            raise NotFoundException("Role set not found")
-        if body.name is not None and body.name != role_set_field.name:
-            role_set_field = await self.role_set_repo.set_name(role_set_field, body.name)
+        server_field = await self.server_repo.get_by_id(server_id)
+        if server_field is None:
+            raise NotFoundException("Server not found")
+        role_set_field = server_field.role_set
+
+        if server_field.role_set_id != role_set_id or role_set_field is None:
+            raise NotFoundException("Role set not found for server")
+        if name is not None and name != role_set_field.name:
+            role_set_field = await self.role_set_repo.set_name(role_set_field, name)
         return role_set_field
 
     async def create_role(
-        self,
-        role_set_id: UUID,
-        body: GameRoleItemCreateRequest, # TODO : Replace body
-        permission_mask: int = 0,
+            self,
+            role_set_id: UUID,
+            server_id: UUID,
+            name: str,
+            min_in_team: int,
+            max_in_team: int,
+            icon_id: UUID | None = None,
+            hidden: bool = False,
+            permission_mask: int = 0,
     ) -> GameRole:
-        # TODO: check if belongs to issuer member server
         if not PERMISSION.check_permission(permission_mask, PERMISSION.EDIT_ROLE_SET):
             raise ForbiddenException("Unable to edit role set")
-        role_set_field = await self.role_set_repo.get_by_id(role_set_id)
-        if not role_set_field:
+        server_field = await self.server_repo.get_by_id(server_id)
+        if server_field is None:
+            raise NotFoundException("Server not found")
+        role_set_field = server_field.role_set
+        if server_field.role_set_id != role_set_id or not role_set_field:
             raise NotFoundException("Role set not found for server")
-
-        icon_url = None
-        icon_id = None
-        # if icon is not None:
-        # TODO: upload icon to minio
-
         try:
             role = await self.role_repo.create(
-                name=body.name,
+                name=name,
                 role_set_id=role_set_id,
-                min_in_team=body.min_in_team,
-                max_in_team=body.max_in_team,
+                min_in_team=min_in_team,
+                max_in_team=max_in_team,
                 icon_id=icon_id,
-                hidden=body.hidden,
+                hidden=hidden,
             )
         except IntegrityForeignException as exc:
             raise NotFoundException(exc.message)
@@ -80,57 +79,43 @@ class GameRoleService:
         return role
 
     async def update_role(
-        self,
-        role_id: UUID,
-        body: GameRoleItemUpdateRequest, # TODO : Replace body
-        permission_mask: int = 0,
+            self,
+            role_id: UUID,
+            server_id: UUID,
+            name: str | None = None,
+            min_in_team: int | None = None,
+            max_in_team: int | None = None,
+            hidden: bool | None = None,
+            permission_mask: int = 0,
     ) -> GameRole:
-        # TODO: check if belongs to issuer member server
         if not PERMISSION.check_permission(permission_mask, PERMISSION.EDIT_ROLE_SET):
             raise ForbiddenException("Unable to edit role set")
+        server_field = await self.server_repo.get_by_id(server_id)
+        if server_field is None:
+            raise NotFoundException("Server not found")
         role = await self.role_repo.get_by_id(role_id)
-        if not role:
-            raise NotFoundException("Role not found")
+        if not role or server_field.role_set_id != role.role_set_id:
+            raise NotFoundException("Role not found for server")
 
-        if body.name is not None and body.name != role.name:
-            role = await self.role_repo.set_name(role, body.name)
-        if body.hidden is not None and body.hidden != role.hidden:
-            role = await self.role_repo.set_hidden(role, body.hidden)
-        if body.min_in_team is not None:
-            role = await self.role_repo.set_min(role, body.min_in_team)
-        if body.max_in_team is not None:
-            role = await self.role_repo.set_max(role, body.max_in_team)
+        if name is not None and name != role.name:
+            role = await self.role_repo.set_name(role, name)
+        if hidden is not None and hidden != role.hidden:
+            role = await self.role_repo.set_hidden(role, hidden)
+        if min_in_team is not None:
+            role = await self.role_repo.set_min(role, min_in_team)
+        if max_in_team is not None:
+            role = await self.role_repo.set_max(role, max_in_team)
         return role
 
-    async def delete_role(self, role_id: UUID, permission_mask: int = 0) -> None:
-        # TODO: check if belongs to issuer member server
+    async def delete_role(self, role_id: UUID, server_id: UUID, permission_mask: int = 0) -> None:
         if not PERMISSION.check_permission(permission_mask, PERMISSION.EDIT_ROLE_SET):
             raise ForbiddenException("Unable to delete role")
+        server_field = await self.server_repo.get_by_id(server_id)
+        role = await self.role_repo.get_by_id(role_id)
+        if server_field is None:
+            raise NotFoundException("Server not found for server")
+        if not role or server_field.role_set_id != role.role_set_id:
+            raise NotFoundException("Role not found for server")
         deleted = await self.role_repo.delete(role_id)
         if not deleted:
             raise NotFoundException("Role not found")
-
-    # async def update_role_icon(
-    #     self, server_id: UUID, role_id: UUID, role_set_id: UUID, icon: UploadFile | None, permission_mask: int = 0
-    # ) -> GameRole:
-    #     server = await self.server_repo.get_by_id(server_id)
-    #     if not server:
-    #         raise NotFoundException("Server not found")
-    #     role = await self.role_repo.get_by_id(role_id)
-    #     if not role or role.role_set_id != role_set_id:
-    #         raise NotFoundException("Role not found")
-    #     if not PERMISSION.check_permission(permission_mask, PERMISSION.EDIT_ROLE_SET):
-    #         raise ForbiddenException("Unable to edit role icon")
-    #
-    #     if icon is None:
-    #         # remove icon
-    #         role = await self.role_repo.set_icon(role, None, None)
-    #         return role
-    #
-    #     content = await icon.read()
-    #     filename = icon.filename
-    #     prefix = f"role_icons/{role_set_id}"
-    #     await minio_manager.upload_file_object(prefix, filename, content)
-    #     icon_url = f"{minio_manager.endpoint}/{minio_manager.bucket_name}/{prefix}/{filename}"
-    #     role = await self.role_repo.set_icon(role, icon_url, None)
-    #     return role
